+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `widget::Indicator` to draw a single indicator
 - Rename `widget::RenderIndicator` to `widget::OrIndicators`
 - Change Highlight Symbol to actually be drawn (fixes #1)
+  - The Highlight Symbol draw width is configurable
 
 ## 0.1.0